+++ src/frontend/js/utils/test/factories/joanie.ts
@@ -34,6 +34,7 @@ import {
   CertificateOrderWithPaymentInfo,
   CredentialOrderWithPaymentInfo,
   EnrollmentLight,
+  OrderGroup,
 } from 'types/Joanie';
 import { CourseStateFactory } from 'utils/test/factories/richie';
 import { FactoryHelper } from 'utils/test/factories/helper';
@@ -229,6 +230,25 @@ export const CourseLightFactory = factory((): CourseLight => {
   };
 });
 
+export const OrderGroupFactory = factory((): OrderGroup => {
+  const seats = faker.number.int({ min: 5, max: 100 });
+  return {
+    id: faker.string.uuid(),
+    is_active: true,
+    nb_seats: seats,
+    nb_available_seats: faker.number.int({ min: 2, max: seats }),
+  };
+});
+
+export const OrderGroupFullFactory = factory((): OrderGroup => {
+  return {
+    id: faker.string.uuid(),
+    is_active: true,
+    nb_seats: faker.number.int({ min: 5, max: 100 }),
+    nb_available_seats: 0,
+  };
+});
+
 export const CourseProductRelationFactory = factory((): CourseProductRelation => {
   return {
     id: faker.string.uuid(),
@@ -236,6 +256,7 @@ export const CourseProductRelationFactory = factory((): CourseProductRelation =>
     course: CourseFactory().one(),
     product: ProductFactory().one(),
     organizations: OrganizationFactory().many(1),
+    order_groups: [],
   };
 });